remove mutex copying
Copying a Logger's embedded mutex is not allowed according to the
documentation, and trips the race detector.

This rewrites the mutex to only protect the io.Writer, and in such a way
that it can be copied. The result no longer trips the race detector, but
may have surprising semantics.